Moved VersionControlJob to .util package - (Task #539)
---
Task #539: Refactor SVN backend
